pyproject.toml
[project]
name = "verify-test"
version = "0.1.0"
description = ""
authors = [
    {name = "lilpacy",email = "revivedtomorrow@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3 (>=1.40.69,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-crc32c (>=1.7.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
